Register views, viewmodels and UI. Still needs address lookup and SpecFlow tests to be added for validation

+++ src/SFA.Apprenticeships.Web.Candidate/ViewModels/DateViewModel.cs
@@ -2,10 +2,10 @@
 {
     public class DateViewModel
     {
-        public short Day { get; set; }
+        public int Day { get; set; }
 
-        public short Month { get; set; }
+        public int Month { get; set; }
 
-        public short Year { get; set; }
+        public int Year { get; set; }
     }
 }
